refactor(GN-43): useImageManager에서 렌더링까지 맡도록 변경
