+++ src/hooks/useImageManager.tsx
@@ -1,0 +1,33 @@
+import { useState } from 'react';
+
+import ImageCard from '@/components/MyActivityPage/ImageCard';
+import ImageInput from '@/components/MyActivityPage/ImageInput';
+
+// 이미지 리스트 관리를 돕는 훅 (추가 / 확인 / 제거)
+const useImageManager = (maxLength: number) => {
+  const [imageFiles, setImageFiles] = useState<File[]>([]);
+
+  const handleAddImage = (file: File) => {
+    setImageFiles((prev) => [...prev, file]);
+  };
+
+  const handleDeleteImage = (file: File) => {
+    setImageFiles((prev) => prev.filter((f) => f !== file));
+  };
+
+  const renderImageManager = () => (
+    <ul className="flex w-full flex-wrap gap-2 pc:gap-6 tablet:gap-4">
+      <ImageInput
+        disabled={imageFiles.length >= maxLength}
+        onChange={handleAddImage}
+      />
+      {imageFiles.map((file) => (
+        <ImageCard image={file} onClickDelete={handleDeleteImage} />
+      ))}
+    </ul>
+  );
+
+  return { imageFiles, renderImageManager };
+};
+
+export default useImageManager;
